% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demodulate.R
\name{demodulate}
\alias{demodulate}
\title{demodulate}
\usage{
demodulate(x, p)
}
\arguments{
\item{x}{Data. It needs to be in vector format.}

\item{p}{Dominant periodicity of the data.}
}
\value{
Returns a list of three objects: the input data, the amplitude, and the phase.
}
\description{
Function to apply complex demodulation, a mathematical time series analysis method, to seasonal infectious disease data.
}
\details{
The method of complex demodulation is in three parts. First, the data is multiplied by a demodulating function. Second, a filter is applied (here, a centred moving average). Third, the amplitude and phase are extracted from the filtered, demodulated data. The output is a list of two objects: the amplitude, and the phase.
}
\examples{
x <- synthdata(seq(1,120), 12, 0.2, 0.2)
demodulate(x, 12)

}
\references{
Hogan, A.B., Glass, K. and Anderssen, R.S. (2017) Complex demodulatuion: a novel #' time series method for analysing infectious diseases. \emph{The ANZIAM Journal} \strong{59}(1) 51-60.
}
\seealso{
\code{\link{synthdata}}
}
