% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdata.R
\name{synthdata}
\alias{synthdata}
\title{synthdata}
\usage{
synthdata(t, p, a = 0, b = 0)
}
\arguments{
\item{t}{Number of time points for which to generate data}

\item{p}{Period of the dominant annual cycle}

\item{a}{Degree of alterating peak sizes, in the range [0,1]}

\item{b}{Degree of delay in alternating peaks, in the range [0,1]}
}
\value{
The output will be a synthetic timeseries of length t.
}
\description{
Generates synthetic biennial epidemic data.
}
\examples{
synthdata(seq(1,120), 12, 0.3, 0.2)

}
\references{
Hogan, A.B., Glass, K. and Anderssen, R.S. (2017) Complex demodulatuion: a novel #' time series method for analysing infectious diseases. \emph{The ANZIAM Journal} \strong{59}(1) 51-60.
}
\seealso{
\code{\link{demodulate}}
}
